+++ Sources/Sentry/SentrySessionTracker.m
@@ -1,4 +1,5 @@
 #import "SentrySessionTracker.h"
+#import "SentryClient+Private.h"
 #import "SentryClient.h"
 #import "SentryFileManager.h"
 #import "SentryHub.h"